Had to ignore error id 131220, kept insisting that my fragment code isn't setup for integer framebuffers, but it works and I can't figure what it's complaining about

+++ src/shaders/BasicGOL/GeoMesh.fs
@@ -4,5 +4,5 @@
 layout(location=0) out uint lifeBoolOut;
 
 void main() {
-    lifeBoolOut = 1;
+    lifeBoolOut = uint(1);
 }
